refactor: change pages to app
