+++ src/app/layout.tsx
@@ -1,6 +1,5 @@
 import '../styles/index.css';
 import { ThemeProvider } from '../ThemeContext';
-import ErrorBoundary from '../components/ErrorBoundary';
 import MainLayout from '../layouts/main';
 import type { ReactNode } from 'react';
 
@@ -8,11 +7,9 @@ const Layout = ({ children }: { children: ReactNode }) => {
   return (
     <html>
       <body>
-        <ErrorBoundary fallback={<p>ErrorBoundary: Something went wrong.</p>}>
-          <ThemeProvider>
-            <MainLayout>{children}</MainLayout>
-          </ThemeProvider>
-        </ErrorBoundary>
+        <ThemeProvider>
+          <MainLayout>{children}</MainLayout>
+        </ThemeProvider>
       </body>
     </html>
   );
