+++ src/app/not-found.tsx
@@ -1,0 +1,5 @@
+const NotFound = () => {
+  return <div>Page not found 404</div>;
+};
+
+export default NotFound;
