--- src/pages/404.tsx
@@ -1,5 +1,0 @@
-const Page404 = () => {
-  return <div>Page not found 404</div>;
-};
-
-export default Page404;
