--- src/pages/index.tsx
@@ -1,5 +1,0 @@
-const MainPage = () => {
-  return '';
-};
-
-export default MainPage;
